fastAPI/requirements.txt
# FastAPI and server
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Document processing
PyMuPDF>=1.19.0  # PDF processing
Pillow>=8.3.0    # Image processing
pytesseract>=0.3.8  # OCR

# LLM and API clients
langchain>=0.1.0
openai>=1.0.0
anthropic>=0.3.0
google-cloud-aiplatform>=1.25.0

# Authentication
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4  # Password hashing
email-validator>=1.1.3  # Email validation

# MongoDB
motor>=3.3.0  # Async MongoDB driver
pymongo>=4.5.0
beanie>=1.21.0  # MongoDB ODM

# Utilities
python-dotenv>=0.19.0
pydantic>=1.8.0
requests>=2.26.0
aiohttp>=3.8.0
tenacity>=8.0.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.23.0 